add charactor param

+++ TOVR/Util.cs
@@ -8,11 +8,11 @@ namespace TOVR
 {
 	class Util
 	{
-		public static void WriteNumber(uint address, uint size, uint value, uint min, uint max)
+		public static void WriteNumber(uint address, uint size, long value, uint min, uint max)
 		{
 			if (value < min) value = min;
 			if (value > max) value = max;
-			SaveData.Instance().WriteNumber(address, size, value);
+			SaveData.Instance().WriteNumber(address, size, (uint)value);
 		}
 	}
 }
